refactoring and created auth feture module
