+++ src/app/app-routing.module.ts
@@ -2,8 +2,7 @@ import { NgModule } from '@angular/core';
 
 import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
-import { LoginComponent } from './login/login.component';
-import { RegisterComponent } from './register/register.component';
+
 const routes: Routes = [
         {
           path:'',// this for if this is empty default 
@@ -21,13 +20,19 @@ const routes: Routes = [
           loadChildren:  () => import('./products/products.module').then(m => m.ProductsModule)
         },
         {
-          path:'login',
-          component:LoginComponent
-        },
-        {
-          path:'register',
-          component: RegisterComponent
+          path: 'auth',
+          loadChildren: () => import('./auth/auth.module').then(m => m.AuthModule) //'./auth/auth.module#AuthModule' 
         }
+
+        // to refactoring we are gonna move on the AuthRouting module
+        // {
+        //   path:'login',
+        //   component:LoginComponent
+        // },
+        // {
+        //   path:'register',
+        //   component: RegisterComponent
+        // }
   ];
   
   @NgModule({
