+++ src/app/auth/auth.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { AuthRoutingModule } from './auth-routing.module';
+import { LoginComponent } from './login/login.component';
+import { RegisterComponent } from './register/register.component';
+import { SharedModule } from '../shared/shared.module';
+
+
+@NgModule({
+  declarations: [
+    LoginComponent, RegisterComponent
+  ],
+  imports: [
+    CommonModule,
+    AuthRoutingModule,
+    SharedModule
+  ]
+})
+export class AuthModule { }
