+++ src/app/auth/login/login.component.html
@@ -42,7 +42,7 @@
         <button mat-raised-button color="primary" 
          (click)="login()">Login</button>
          <span>
-             If you dont have an account? <a [routerLink]="['/register']"> register</a> here.
+             If you dont have an account? <a [routerLink]="['/auth/register']"> register</a> here.
          </span>
     </mat-card-actions>
 </mat-card>
