+++ src/app/auth/register/register.component.html
@@ -67,7 +67,7 @@
         <button mat-raised-button color="primary" 
         (click)="register()">Register</button>
         <span class="left-spacer">Already have a account login here 
-            <a [routerLink]="['/login']">login</a>
+            <a [routerLink]="['/auth']">login</a>
             here
         </span>
     </mat-card-actions>
