+++ src/app/auth/register/register.component.scss
